Update README.md
Changed the name of the project from GoFAST to GoPAST. Unfortunately due to limited time I can no longer release config files or shell scripts for both FreeBSD and Debian, so concentrating on developing GoPAST on Debian for now. Hopefully in the future will continue development of GoPAST on FreeBSD again as it is an amazing operating system.